sqe: change offset to u64
An offset is a file position, so it can be bigger than usize in
32-bit platforms.

+++ src/sqe.rs
@@ -221,7 +221,7 @@ impl<'a> SubmissionQueueEvent<'a> {
         &mut self,
         fd: RawFd,
         buf: &mut [u8],
-        offset: usize,
+        offset: u64,
         buf_index: usize,
     ) {
         let len = buf.len();
@@ -283,7 +283,7 @@ impl<'a> SubmissionQueueEvent<'a> {
         &mut self,
         fd: RawFd,
         buf: &[u8],
-        offset: usize,
+        offset: u64,
         buf_index: usize,
     ) {
         let len = buf.len();
